Revert "Reapply "Feat(Stats): Fixed enemy and tower stats/display""
This reverts commit 55ea263e389ff986c5132d76c522dc8d810805d9.

+++ src/entities/towers/Tower.cpp
@@ -5,7 +5,7 @@
 
 std::vector<Tower*> Tower::allInstances;
 
-Tower::Tower(Vector2 position, int level, const std::string& path, int damage, double speed, int range, int spAttackRegenerationTime)
+Tower::Tower(Vector2 position, int level, const std::string& path, int damage, int speed, int range, int spAttackRegenerationTime)
     : position(position), level(level), texturePath(path), damage(damage), speed(speed), range(range),
        spAttackRegenerationTime(spAttackRegenerationTime)
 {
